requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
web3>=5.0.0
ipfsapi>=0.4.4
